// extensions/codegrepper/src/interface.tsx
export interface Answer {
  id: number;
  search_id?: any;
  answer: string;
  up_votes: number;
  down_votes: number;
  user_id: number;
  nickname?: any;
  language: string;
  created_at: string;
  updated_at: string;
  bounty?: any;
  bounty_approved?: any;
  solution_urls: string;
  source_id: number;
  is_deleted: number;
  source_url: string;
  do_hide: number;
  do_hide_r?: any;
  do_hide2: number;
  do_hide2_r?: any;
  is_private: number;
  is_widget: number;
  video_name: string;
  t_upvotes: number;
  t_upvotes_d: number;
  t_downvotes: number;
  t_downvotes_d: number;
  t_copies: number;
  t_follow_up_searches: number;
  t_no_follow_up_searches: number;
  t_other_answer_clicked: number;
  t_score: string;
  did_help_per: string;
  t_total_results: number;
  t_users_count: number;
  cluster_id: string;
  cluster_id_big: string;
  desp_for_ans: number;
  lang_prob?: any;
  humor_prob: string;
  new_sa_count: number;
  t_scorecat1?: any;
  t_ttr_cat1?: any;
  t_scorecat2?: any;
  t_ttr_cat2?: any;
  t_scorecat3?: any;
  t_ttr_cat3?: any;
  t_scorecat4?: any;
  t_ttr_cat4?: any;
  t_scorecat5?: any;
  t_ttr_cat5?: any;
  t_comments: number;
  answer_md5: string;
  problem_cluster_id: number;
  fun_name: string;
  profile_slug: string;
  search_answer_id: number;
  donate_link: string;
  term: string;
  raw_score: string;
  score: string;
  total_results: number;
  upvotes: number;
  downvotes: number;
  i_upvoted?: any;
  i_downvoted?: any;
  is_my_answer: number;
  search_answer_result_id: string;
}

export interface MoreAnswer {
  id: number;
  search_id?: any;
  answer: string;
  up_votes: number;
  down_votes: number;
  user_id: number;
  nickname?: any;
  language: string;
  created_at: string;
  updated_at: string;
  bounty?: any;
  bounty_approved?: any;
  solution_urls: string;
  source_id: number;
  is_deleted: number;
  source_url: string;
  do_hide: number;
  do_hide_r?: any;
  do_hide2: number;
  do_hide2_r?: number;
  is_private: number;
  is_widget?: number;
  video_name: string;
  t_upvotes: number;
  t_upvotes_d: number;
  t_downvotes: number;
  t_downvotes_d: number;
  t_copies: number;
  t_follow_up_searches: number;
  t_no_follow_up_searches: number;
  t_other_answer_clicked: number;
  t_score: string;
  did_help_per: string;
  t_total_results: number;
  t_users_count: number;
  cluster_id: string;
  cluster_id_big: string;
  desp_for_ans: number;
  lang_prob: string;
  humor_prob: string;
  new_sa_count: number;
  t_scorecat1: string;
  t_ttr_cat1?: number;
  t_scorecat2: string;
  t_ttr_cat2?: number;
  t_scorecat3: string;
  t_ttr_cat3?: number;
  t_scorecat4: string;
  t_ttr_cat4?: number;
  t_scorecat5: string;
  t_ttr_cat5?: number;
  t_comments: number;
  answer_md5: string;
  problem_cluster_id?: number;
  fun_name: string;
  profile_slug: string;
  search_answer_id: number;
  donate_link: string;
  term: string;
  raw_score: string;
  score: string;
  total_results: number;
  upvotes: number;
  downvotes: number;
  i_upvoted?: any;
  i_downvoted?: any;
  is_my_answer: number;
}

export interface CodeGrepperObject {
  answers: Answer[];
  products: any[];
  writeups: any[];
  language_guess: string;
  f_more_results: number;
  grepcc_tips_left: number;
  allow_wrong_answers_feedback: number;
  more_answers: MoreAnswer[];
}